+++ e3nn_jax/_irreps.py
@@ -198,7 +198,7 @@ class Irrep:
         yield self.p
 
     def __lt__(self, other):
-        return (self.l, self.p) < (other.l, other.p)
+        return (self.l, -self.p * (-1) ** self.l) < (other.l, -other.p * (-1) ** other.l)
 
 
 jax.tree_util.register_pytree_node(Irrep, lambda ir: ((), ir), lambda ir, _: ir)