+++ ChangeLog.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- **breaking change** change the ordering of `Irrep`. Now it matches with `Irrep.iterator`.
 
 ## [0.7.0] - 2022-08-03
 ### Changed